requirements.txt
Django>=4.0,<5.0
djangorestframework
gunicorn
Pillow
numpy
tensorflow==2.18.0
whitenoise
dj-database-url
